put integration test stuff back where it belongs
