+++ linkerd/app/integration/src/data/foo-ns1/csr.der
