+++ linkerd/app/integration/src/data/foo-ns1/key.p8
