+++ linkerd/app/integration/src/data/foo-ns1/token.txt
