+++ linkerd/app/integration/src/data/gen-certs.sh
